requirements.txt
discord.py
pymongo
Pillow
python-dotenv
easy-pil
